requirements.txt
discord
asyncio
python-dotenv==0.21.0
colorama